fix writingFile and header for better file handling

+++ scraping/lapor_scraping.py
@@ -7,7 +7,13 @@ def get_url(query, page):
     """
     Generate URL from query and page number.
     """
-    url = 'https://www.lapor.go.id/search?q={}&page={}'.format(query, page)
+    queryLength = len(query.split())
+    if queryLength == 1:
+        url = 'https://www.lapor.go.id/search?q={}&page={}'.format(query, page)
+    else:
+        query = query.split()
+        param = '+'.join(query)
+        url = 'https://www.lapor.go.id/search?q={}&page={}'.format(param, page)
     return url
   
 def get_html_source(url):
@@ -88,9 +94,25 @@ def writeFile(dataframe):
     """
     Save the DataFrame into a csv file.
     """
-    print('   + Writing file . .')
-    with open('{}.csv'.format(query), 'a+') as file:
-        dataframe.to_csv(file, header=True)
+    DATA_PATH = 'data/'
+    QUERY_PATH = DATA_PATH + '{}.csv'.format(QUERY)
+
+    # if csv file of query is exist
+    if os.path.exists(QUERY_PATH):
+        # open the file and will append the new data to the file
+        with open(QUERY_PATH, 'a+') as file:
+            dataframe.to_csv(file, header=False)
+    
+    # if data folder doesn't exist
+    elif os.path.exists(DATA_PATH) == False:
+        # create folder for data
+        os.mkdir(DATA_PATH)
+
+    # if csv file of query doesn't exist
+    else:
+        # write a new file inside the data folder
+        with open(QUERY_PATH, 'w') as file:
+            dataframe.to_csv(file, header=True)
   
 def main():   
     global QUERY, PAGE_LEN
